Added Rudimentary Headlines Feature
Added a new Headline model with the intention of the user being able to select
certain tags for which a new/trending article will display on the front page for
that tag. Like a newspaper. Currently, it just updates the headline to the
newest article for the last clicked feed with that tag.

For one, it would at least be good to check by date. More importantly, feeds
need to be checked periodically to update what's new for each category. This
can also provide a list of overall newest articles.

Eventually, using outside sources to determine article popularity would be ideal.
Comments from those outside sources could also be linked to. And perhaps users
could even choose which sources they want to use to determine popularity.